refactor: simplify public API and improve method naming - Change core V2 auth methods to internal - Rename login to loginWithPasskey for clarity - Move EVMTransaction.toMetaMaskFormat to MetaMaskConnector - Rename handleVerification to handleEmailVerification
